requirements.txt
Flask
python-dotenv
sqlalchemy 
flask-sqlalchemy
psycopg2-binary